pyproject.toml
[project]
name = "megaschool"
version = "0.1.0"
description = ""
authors = [
    {name = "korsNaike",email = "kirillgrinev2003@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "dependency-injector (>=4.45.0,<5.0.0)",
    "openai (>=1.60.2,<2.0.0)",
    "tenacity (>=9.0.0,<10.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "colorlog (>=6.9.0,<7.0.0)",
    "googlesearch-python (>=1.3.0,<2.0.0)",
    "faker (>=35.2.0,<36.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.3"
anyio = "^4.8.0"
asgi-lifespan = "^2.1.0"

